Update comment from Yahoo Weather to Night Shift
With Yahoo Weather having been removed and the readme already updated to reflect that, Yahoo is still referenced in the top of the script in a comment. Remove this last reference to Yahoo Weather and replace it with Night Shift, to avoid future people like me being confused as to where Yahoo is involved!